+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 📋 Changelog for "LOOPIS Content"
 
+## 0.37 (2026-08-25)
+- Former CPT 'forum' renamed to 'news'
+- Former CPT 'support' deprecated and replaced by 'forum'
+
 ## 0.36 (2026-05-13)
 - Support for multisite: separate configuration for HQ & Local. (WIP)